adds a time limit to the SpinSpinnerToColor and adds comments describing each of the Spinner commands
This commit does two things. The first thing it does is it adds comments to describe each of the Spinner commands. It also times how long the spinner stays on a color in the SpinSpinnerToColor command. If the color does not change in a certain amount of time it aborts the command.