updates turret angle constants from testing
This change updates the turret angle constants
from testing done Tuesday. Note that PID values
were not updated in this commit. The constant
VOLTS_PER_DEGREE was changed to DEGREES_PER_VOLT
to correctly reflect the units from the angle
calculation. The speed of the turret motor was
added to the Turret shuffleboard tab to enable
monitoring whether the PID command tells the
turret to spin in the correct direction.