updates the vision python application
This commmit updates the vision python application. It allows the color
filtration values to be modified in shuffleBoard. Also, it upgrades
the code that draws the bounding box by allowing only the largest
bounding box available to be used. This prevents any small anomalies off
to the side from being detected.