adds the PIDVisionShoot command
This commit adds the PIDVisionShoot command to the Shooter subsystem, along with all of the
accompanying changes. For example, a getZone() method was added to the Vision subsystem
that uses the distance from the vision python program to estimate the zone the robot is
currently in. Also, a getZoneShooterSpeed() method was added to the Shooter subsystem that
uses the zone estimated by the getZone() method to determine the speed necessary for the
shooter flywheel. These methods are used in the PIDVisionShoot command to shoot the balls
at a speed determined by the current zone. Finally, this commit adds several constants
to the Constants class that are usefull for the PIDVisionShoot command.
This commit does NOT add the correct flywheel speeds or functionality in the
PIDVisionShoot command to raise and lower the shooter.