adds the VisionShootBallsAuto command
This commit adds the VisionShootBallsAuto command and
all of the associated commands. It is a sequential command that runs
VisionFindTarget() and PIDVisionFindTarget. Then it runs the
ParallelDeadlineVisionShoot() command, which is a ParallelDeadline
command that ends when one ball is fired. The two commands it runs are the
TargetAndFire() command and the PIDVisionShoot() command. The TargetAndFire()
command turns the turret with PIDVisionTurretTarget() and the shoots the
balls with FeedBalls(). One thing to note with this commit is that only the
VisionShootBallsAuto() command is added to the testing dashboard. The
other commands can be registered in the future for testing purposes.